Remove dockerhub login step and use ECR public gallery ubuntu image (#78)
This is required so that we do not need to support dockerhub and its credentials. 
1. Remove dockerhub login step from the codebuild spec file.  
2. Use [ECR public gallery ubuntu image](https://gallery.ecr.aws/ubuntu/ubuntu) in the Canary Docker Image instead

### Testing
For testing I added this commit on top of the `canary` branch and pushed to a new upstream branch [`docker-on-ecr-on-canary-branch`](https://github.com/aws-controllers-k8s/sagemaker-controller/tree/docker-on-ecr-on-canary-branch). Then ran the canary test job on this branch.