Notebook update code + updated integration test (#61)
Notebook Instance Update PR

This is a PR that adds the update operation to the notebook operator and also contains a sample.

How update works:

1. Controller stops the Notebook if its not Stopped already.(Sets a status field "StoppedByControllerMETA") if it did stop the notebook.
2. Controller requeues during intermediate "Stopping" state until it reaches "Stopped"
3. Controller updates the notebook and also sets the StoppedByControllerMETA to "UpdateTriggered" if its not nil.
4. After the update the controller starts the Notebook if StoppedByControllerMETA is "UpdateTriggered".
5. The controller resets the StoppedByControllerMETA status field.

Note:

The update operation is supposed to start the notebook after an update if it was not in a stopped state and keep the notebook in the stopped state if it was in the stopped state prior to an update.

We use a status field StoppedByControllerMETA to keep track of whether the notebook was stopped by a controller and an annotation done_updating to keep track of whether an update has finished. 

Why is StoppedByControllerMETA a status and not an annotation?
After the controller stops a Notebook Instance, it can’t update the notebook because it will usually be in the stopping state and will stay in “Stopping” for at least a minute. Because of this an error is returned as the notebook can only be updated when it is in the stopped state. If we dont return an error the spec gets updated and the controller wont be able to detect a delta.
Currently it is impossible for the controller to update an annotation if an update call has failed. If an update fails the controller patches the resource status and not the metadata(https://github.com/aws-controllers-k8s/runtime/blob/055b089c6a508317ad4bb57ce037dc55061e1829/pkg/runtime/reconciler.go#L235). We can’t also not return an error because that would patch the spec too(which would end up with the controller not calling update).

Note:
Even after 0.7.1, you cant update annotations(metadata) without also updating the spec, we cannot update the spec when we stop the notebook as that would prevent the update from occurring(runtime will not detect a diff).(https://github.com/aws-controllers-k8s/runtime/blob/d7b6e45c7ba9e54664b49f681747abf275a232f9/pkg/runtime/reconciler.go#L587)

Files added:

custom_sdk_api.go: Handles api calls to Sagemaker

Hooks.go: Contains code to requeue and also customsetoutputDescribe which is used to set the resource synced conditions and also used to start the Notebook after an update(if it was not stopped before the update).


notebook_instance/sdk_read_one_post_set_output.go.tpl:
Calls custom_set_describe
notebook_instance/sdk_read_one_pre_set_output.go.tpl. 
sdk_read_one_pre_set_output.go.tpl sets the LifecycleConfigName spec field(code generator does not generate it).

notebook_instance/sdk_update_pre_build_request.go.tpl:
Requeues if in the stopping/updating/pending state. If the Notebook is in the InService state stopNotebookInstance is called, if it does not return an error, the status field StoppedByControllerMETA is set to true.

notebook_instance/sdk_update_post_set_output.go.tpl:
Sets the is_updating status and sets the resource synced condition to false so the controller requeues.
Covered in Create/Delete:
https://github.com/aws-controllers-k8s/sagemaker-controller/pull/56

Removed a check for assert for UpdateTriggered because it makes the test much more flakey because update can take as little as a minute at times

————————————————————————————————————

Custom code:
pkg/resource/notebook_instance/custom_set_delta.go : Fills in default values to prevent an update after creation
pkg/resource/notebook_instance/hooks.go :  Contains code to requeue and set sync conditions, also contains custom_set_output_describe that starts a Notebook(after an update).
templates/notebook_instance/sdk_delete_pre_build_request.go.tpl : Stops the notebook and re queues on any state other than Stopped/Failed.