requeue until Stopped for job style resources (#88)
Description of changes:
- requeue until Stopped for job style resources
  - integ test changes for the same
- Reduce terminal codes to only those which can be fixed by a change in spec. Reason being the latest runtime 0.13+ ignores metadata and status updates leaving customer with no way to get the controller reconcile if the terminal code is not related to spec field change. I will make this change for other resources in a separate PR
- consolidate hook related code to one file and refractor to use common pkg to reduce code duplication
  - removed post `sdk_create_post_set_output` hook because runtime does a readone and paches the status post create

TODO:
During testing, I found that if there is an error during delete it is not reflected in the resource conditions. Will put a PR to codegen and create a new PR with those changes

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.