Use tmp file in unit-tests to allow for testing portability (#85)
Description of changes:
Unit tests fail when not using GNU sed this is because command uses `sed -i` which when not using GNU sed needs to specify a string afterwards which can be defaulted to `sed -i ''` However when attempting this allows for the unit-tests to pass but it created tmp files on my end with suffix ''. Ended up creating .tmp files initially and then removing them, this is consistent with the GNU internal implementation. https://www.gnu.org/software/sed/manual/sed.html#index-_002di 

> GNU sed does this by creating a temporary file and sending output to this file rather than to the standard output

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.