terminal codes clean up (#89)
Description of changes:

Be conservative in the list of terminal codes. Specify only those for which the user needs to change a value in spec.
  - looked at each resource's APIs errors and removed the ones that are not thrown in any of the CRUD APIs

Long pending backlog item

Testing:
will rely on PR Build

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.