Update iam passrole readme instructions (#130)
Description of changes:
Previously the policy provided in the readme was allowing for iamPassRole to have access to all resources. 
This changes this so that it is only scoped to a resource in the account the controller was created in.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.