Fix late initialization field (#165)
Issue #, if available:
When a distributed/parallelized pytorch workload is run, AlgorithmSpecification.MetricDefinitions does not late initialize which makes the controller re queue for a long time. 

Description of changes:
Edited generator.yaml to remove late_initialize and to ignore comparisons(ACK trainingjob does not support update operations so this wont effect the user). 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.