Handle server side default for inputDataConfig in TrainingJob (#168)
Bug fix for unhandled case in training job. `inputDataConfig` is an optional field. Its default value is an empty list. Without this hook, the controller reconciles because of diff between null(desired) and empty list(observed) when `inputDataConfig` is not specified
  - usecase: If user bundles the dataset and training code into the training image or downloads dataset at runtime

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.