Refactor e2e deletion and unmark adopted Endpoint Test (#216)
Issue #, if available:

Description of changes:
Disabling the adopted endpoint test as canary since this does not provide value for a canary, will keep as e2e test. Refactors  
```
if k8s.get_resource_exists(reference): 
 _, deleted = k8s.delete_custom_resource(reference, cfg.JOB_DELETE_WAIT_PERIODS, cfg.JOB_DELETE_WAIT_LENGTH) 
> assert deleted 
```
into a shared method, and adds a retry loop for this assert.
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.